--- a/pkg/ccl/storageccl/import.go
+++ b/pkg/ccl/storageccl/import.go
@@ -210,7 +210,7 @@ func evalImport(ctx context.Context, cArgs storage.CommandArgs) (*roachpb.Import
 	log.Infof(ctx, "import [%s,%s)", importStart, importEnd)
 
 	var rows rowCounter
-	var iters []engine.Iterator
+	var iters []engine.SimpleIterator
 	for _, file := range args.Files {
 		if log.V(2) {
 			log.Infof(ctx, "import file [%s,%s) %s", importStart, importEnd, file.Path)
@@ -291,9 +291,16 @@ func evalImport(ctx context.Context, cArgs storage.CommandArgs) (*roachpb.Import
 	g, gCtx := errgroup.WithContext(ctx)
 	startKeyMVCC, endKeyMVCC := engine.MVCCKey{Key: args.DataSpan.Key}, engine.MVCCKey{Key: args.DataSpan.EndKey}
 	iter := engineccl.MakeMultiIterator(iters)
+	defer iter.Close()
 	var keyScratch, valueScratch []byte
-
-	for iter.Seek(startKeyMVCC); iter.Valid() && iter.UnsafeKey().Less(endKeyMVCC); iter.NextKey() {
+	for iter.Seek(startKeyMVCC); ; iter.NextKey() {
+		ok, err := iter.Valid()
+		if err != nil {
+			return nil, err
+		}
+		if !ok || !iter.UnsafeKey().Less(endKeyMVCC) {
+			break
+		}
 		if len(iter.UnsafeValue()) == 0 {
 			// Value is deleted.
 			continue
@@ -304,8 +311,6 @@ func evalImport(ctx context.Context, cArgs storage.CommandArgs) (*roachpb.Import
 		key := engine.MVCCKey{Key: keyScratch, Timestamp: iter.UnsafeKey().Timestamp}
 		value := roachpb.Value{RawBytes: valueScratch}
 
-		var ok bool
-		var err error
 		key.Key, ok, err = kr.RewriteKey(key.Key)
 		if err != nil {
 			return nil, err
@@ -346,9 +351,6 @@ func evalImport(ctx context.Context, cArgs storage.CommandArgs) (*roachpb.Import
 				return nil, err
 			}
 		}
-	}
-	if err := iter.Error(); err != nil {
-		return nil, err
 	}
 	// Flush out the last batch.
 	if batcher.Size() > 0 {