The cgo calls involved in iterating an sstable to rekey it and validate
it in AddSSTable are now showing up as significant in pprof (especially
when run with the in-progress sideloading work). This drastically
improves our benchmarks, see next commit.

This uses the github.com/golang/leveldb package, which has a big
"work-in-progress" warning, but the sstable parsing part seems finished.
At very worst, we can always fork it if we need any fixes.

We still need the old one for now to compute mvcc stats. I'm quite
surprised at how much of a difference computing the stats makes,
commenting it out changes the AddSSTable benchmark radically
(100/1000/10000 are ~15/30/46 MB/s).

```
name                         old time/op    new time/op    delta
AddSSTable/100-8               2.14ms ± 2%    2.17ms ± 2%     ~     (p=0.200 n=4+4)
AddSSTable/1000-8              6.68ms ± 9%    6.91ms ±14%     ~     (p=0.690 n=5+5)
AddSSTable/10000-8             52.8ms ± 3%    50.1ms ± 3%   -5.02%  (p=0.016 n=5+5)
Import/AddSSTable/1-8          2.51ms ± 1%    1.80ms ± 1%  -28.39%  (p=0.008 n=5+5)
Import/AddSSTable/100-8        3.72ms ± 1%    2.84ms ± 2%  -23.78%  (p=0.016 n=4+5)
Import/AddSSTable/10000-8      93.4ms ± 6%    81.3ms ± 4%  -13.05%  (p=0.008 n=5+5)
Import/AddSSTable/100000-8      540ms ± 6%     430ms ± 9%  -20.35%  (p=0.008 n=5+5)
ClusterRestore/AddSSTable-8    6.25µs ± 3%    5.51µs ± 6%  -11.84%  (p=0.008 n=5+5)

name                         old speed      new speed      delta
AddSSTable/100-8             8.71MB/s ± 2%  8.55MB/s ± 2%     ~     (p=0.200 n=4+4)
AddSSTable/1000-8            25.8MB/s ± 8%  25.1MB/s ±13%     ~     (p=0.690 n=5+5)
AddSSTable/10000-8           32.4MB/s ± 3%  34.1MB/s ± 3%   +5.24%  (p=0.016 n=5+5)
Import/AddSSTable/1-8         450kB/s ± 0%   626kB/s ± 1%  +39.11%  (p=0.016 n=4+5)
Import/AddSSTable/100-8      4.66MB/s ±10%  6.29MB/s ± 2%  +34.92%  (p=0.008 n=5+5)
Import/AddSSTable/10000-8    17.5MB/s ± 6%  20.1MB/s ± 4%  +14.96%  (p=0.008 n=5+5)
Import/AddSSTable/100000-8   30.2MB/s ± 6%  38.0MB/s ± 8%  +25.61%  (p=0.008 n=5+5)
ClusterRestore/AddSSTable-8  35.8MB/s ± 3%  40.7MB/s ± 6%  +13.62%  (p=0.008 n=5+5)
```

cc @petermattis in case you want to look at the storage/engine changes