@tschottdorf 2017-06-20
Beautifully self-contained commits! No substantial comments. <img class="emoji" title=":lgtm_strong:" alt=":lgtm_strong:" align="absmiddle" src="https://reviewable.io/lgtm_strong.png" height="20" width="61"/> 

---

Reviewed 4 of 4 files at r1, 5 of 5 files at r2, 2 of 2 files at r3.
Review status: all files reviewed at latest revision, 13 unresolved discussions, some commit checks failed.

---

*[pkg/ccl/storageccl/import.go, line 301 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4sWPVbSeo9xDLaEP6:-Kn4sWPVbSeo9xDLaEP7:b-gf6of0) ([raw file](https://github.com/cockroachdb/cockroach/blob/a03c1ebf7f59441bf2b94314381b3e1b98e16b24/pkg/ccl/storageccl/import.go#L301)):*
> ```Go
> 			break
> 		}
> 		if err != nil {
> ```

Doesn't this need to be checked before you `break`?

---

*[pkg/ccl/storageccl/engineccl/multi_iterator_test.go, line 114 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4syOXQVdNB9TwqQ9y:-Kn4syOXQVdNB9TwqQ9z:bolip50) ([raw file](https://github.com/cockroachdb/cockroach/blob/a03c1ebf7f59441bf2b94314381b3e1b98e16b24/pkg/ccl/storageccl/engineccl/multi_iterator_test.go#L114)):*
> ```Go
> 							break
> 						}
> 						if !testutils.IsError(err, test.errRE) {
> ```

Same here, with `err != nil`  you'll always have `!ok` so don't think you'd ever reach this.

---

*[pkg/ccl/storageccl/engineccl/sst_iterator.go, line 88 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4u0cxEqyoVUw77Vzg:-Kn4u0cxEqyoVUw77Vzh:bxc6xg2) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/ccl/storageccl/engineccl/sst_iterator.go#L88)):*
> ```Go
> 		r.err = r.iter.Close()
> 	}
> 	if err := r.iter.Close(); err != nil && r.err == nil {
> ```

The double-close is not a problem?

---

*[pkg/ccl/storageccl/engineccl/sst_iterator.go, line 164 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4v2o5yEP5S6KRU0fw:-Kn4v2o5yEP5S6KRU0fx:bjwlgo6) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/ccl/storageccl/engineccl/sst_iterator.go#L164)):*
> ```Go
> 		// This should never happen unless there is some sort of corruption of
> 		// the keys.
> 		return bytes.Compare(a, b)
> ```

Yet, you don't want to error out?

---

*[pkg/ccl/storageccl/engineccl/sst_iterator_test.go, line 31 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4wTA4tg0bf78-xLZr:-Kn4wTA4tg0bf78-xLZs:b9kpat) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/ccl/storageccl/engineccl/sst_iterator_test.go#L31)):*
> ```Go
> 		ok, err := iter.Valid()
> 		if err != nil {
> 			t.Fatalf("%+v", err)
> ```

here and below: why not just `t.Fatal(err)`?

---

*[pkg/ccl/storageccl/engineccl/sst_iterator_test.go, line 46 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4woHanKskDoIpwXYg:-Kn4woHbEEd9aXJjV5ar:b639ch6) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/ccl/storageccl/engineccl/sst_iterator_test.go#L46)):*
> ```Go
> 	}
> 	iter.Close()
> 	if _, err := iter.Valid(); err != nil {
> ```

don't wanna check that it returns `false, nil`?

---

*[pkg/ccl/storageccl/engineccl/sst_iterator_test.go, line 68 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4x0g9Mul8Fxk0mNjd:-Kn4x0gATkgfMtRy1c6G:b4rnat7) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/ccl/storageccl/engineccl/sst_iterator_test.go#L68)):*
> ```Go
> 			Key: engine.MVCCKey{
> 				Key:       []byte{i},
> 				Timestamp: hlc.Timestamp{WallTime: int64(i + 2)},
> ```

why `i+2`? Intentionally avoiding `WallTime==0`?

---

*[pkg/ccl/storageccl/engineccl/sst_iterator_test.go, line 75 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4x_pWoPzLgz1FgJ9v:-Kn4x_pXvWFmN_8qLCLb:b-skhhbz) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/ccl/storageccl/engineccl/sst_iterator_test.go#L75)):*
> ```Go
> 			t.Fatalf("%+v", err)
> 		}
> 		kv.Key.Timestamp.WallTime = int32(i + 1)
> ```

I don't understand why it's `i+1` here but `i+2` above. Perhaps some comments would help.

---

*[pkg/ccl/storageccl/engineccl/sst_iterator_test.go, line 126 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4xAN99uRs4QUsboeQ:-Kn4xAN99uRs4QUsboeR:b-nnz5hf) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/ccl/storageccl/engineccl/sst_iterator_test.go#L126)):*
> ```Go
> 	c := cockroachComparer{}
> 	if x := c.Compare(keyAIntent, keyA1); x != -1 {
> 		t.Errorf("expected intent to sort first got: %d", x)
> ```

I don't 100% remember the MVCC encoding, but is this really an intent? I thought we mostly stored the latest version inlined in the metadata key (which includes intents, but still).

---

*[pkg/storage/engine/batch.go, line 250 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4yHm-JJBTwENlcnZx:-Kn4yHm0no2lecMWEWPE:b-qle4ja) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/storage/engine/batch.go#L250)):*
> ```Go
> 	timestampLength := 0
> 	if key.Timestamp != (hlc.Timestamp{}) {
> 		timestampLength = 1 + 8
> ```

Comment on the `1` and `8`.

---

*[pkg/storage/engine/batch.go, line 256 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4yLvscftWyR73KiRT:-Kn4yLvscftWyR73KiRU:b-isjr86) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/storage/engine/batch.go#L256)):*
> ```Go
> 	}
> 
> 	dbKey := make([]byte, len(key.Key)+timestampLength+1)
> ```

Comment on the `+1`.

---

*[pkg/storage/engine/batch.go, line 270 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4zDUmJKLNd9D_IOGn:-Kn4zDUmJKLNd9D_IOGo:b67kowf) ([raw file](https://github.com/cockroachdb/cockroach/blob/d7c7d11d8a89bb7f46e42b19162e1054dd18b9b8/pkg/storage/engine/batch.go#L270)):*
> ```Go
> 		}
> 	}
> 	dbKey[pos] = byte(timestampLength)
> ```

`pos` is now `len(dbKey)-1`, no? Might be more illuminating to call it that (ok, perhaps slower).

---

*[pkg/storage/engine/engine.go, line 34 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-Kn4tFFv8zpWGHVDZrpN:-Kn4tFFv8zpWGHVDZrpO:b-ai0vk8) ([raw file](https://github.com/cockroachdb/cockroach/blob/a03c1ebf7f59441bf2b94314381b3e1b98e16b24/pkg/storage/engine/engine.go#L34)):*
> ```Go
> // engine. SimpleIterator implementations are thread safe unless otherwise
> // noted. SimpleIterator is a subset of the functionality offered by Iterator.
> type SimpleIterator interface {
> ```

Initially thought a good name would be `ForwardIterator`, but then some other methods are also absent, so `SimpleIterator` seems fine.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/16614#-:-Kn4zy1e5qjgm9s6raeF:b-1ng7yn)*
<!-- Sent from Reviewable.io -->